# Speed up didToProfile in large queues