This speeds up the `didToProfile` helper on the queue page if the queue
is large (say over 500 accounts). I’m not quite sure why the current
solution is this slow (up to 3s to rerender) since we should only be
looking at 2.25 million lookups maxium in a queue of n = 1500.

Generating the map initially takes around 2ms and any lookups via
`didToProfile` go down from around 40ms to <0.2ms.